requirements.txt
aiogram
discord
markovify
multitasking
pyTelegramBotAPI
python-dotenv
requests
psutil
PyQt5
PyQt5-Qt5
PyQt5-sip
aiosignal